Create go.php
An example of Nofollow Link:

<a href=”https://inforesult.in/” rel=”nofollow”>inforesult</a>

    Read: WordPress plugin to add Nofollow tag attribute
    How to add Nofollow tag to a link

Dofollow links allow google (all search engines) to follow them and reach our website. Giving us link juice and a backlink. If a webmaster is linking back to you with this link both Search Engine and Humans will be able to follow you. The best way to give someone dofollow love is allowing keyword in the anchor text. This means when you are linking to any website or page, use the targeted keyword as anchor text.

An example of Dofollow Link:

<a href=”https://inforesult.in/”>inforesult</a>